# Conflicts:
#	include/abilities.h
#	include/ch/cheato.h
#	include/ch/chuffycont.h
#	include/ch/glowbo.h
#	include/ch/humba.h
#	include/ch/signpost.h
#	include/common.h
#	include/core1/1E2F380.h
#	include/core2/1E691A0.h
#	include/core2/1E78170.h
#	include/core2/1E7AB30.h
#	include/core2/1E97EF0.h
#	include/core2/1E99980.h
#	include/core2/1EA0690.h
#	include/core2/1EB52C0.h
#	include/core2/1EC8070.h
#	include/core2/1EC9740.h
#	include/core2/1ECB0F0.h
#	include/core2/1ECE0B0.h
#	include/core2/1ED3900.h
#	include/core2/1ED8C80.h
#	include/core2/1EDA900.h
#	include/core2/1EE27C0.h
#	include/core2/1EE5DF0.h
#	include/fx/splash.h
#	include/fx/splash25d.h
#	include/gc/sectionskip.h
#	include/su/baddiedialog.h
#	src/core2/1E66990.c
#	src/core2/1ECE0B0.c
#	src/overlays/ch/boggy/fish.c
#	src/overlays/ch/glowbo.c
#	src/overlays/ch/humba.c
#	src/overlays/ch/jadestatue.c
#	src/overlays/ch/jigsaw/dance.c
#	src/overlays/ch/lagoon/ufoext.c
#	src/overlays/ch/newmumbo.c
#	src/overlays/ch/signpost.c
#	src/overlays/gc/sectionskip.c

+++ include/gl/dbstring.h
@@ -5,7 +5,7 @@
 #include "gl/dialog.h"
 
 u32 *_gldbstring_entrypoint_0(u32* TextAddress, u32 a1);
-u32 *_gldbstring_entrypoint_1(u32* TextAddress, u8 index);
-void _gldbstring_entrypoint_2(u16 a0); // this may return something
+u32 *_gldbstring_entrypoint_1(u32* TextAddress, s16 index);
+void _gldbstring_entrypoint_2(s32 a0); // this may return something
 
 #endif
